requirements.txt
web3==6.11.0
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6
schedule==1.2.0
aiohttp==3.9.1
asyncio==3.4.3